+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## 2026.9.2 - 2026-09-02
 
+### Fixed
+
+- Fix polynomial-redos in readConfig trailing-slash regex (CodeQL alert \#1) ([pm-ado-mih0](https://github.com/unbraind/pm-ado/blob/main/.agents/pm/issues/pm-ado-mih0.toon))
+
 ### Security
 
 - Closed the polynomial expression in the relation parser that a coverage edit had introduced ([pm-ado-muqg](https://github.com/unbraind/pm-ado/blob/main/.agents/pm/issues/pm-ado-muqg.toon))